+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerSelectedDate.test.jsx
@@ -6,20 +6,9 @@ class DatePickerDefault extends React.Component {
   constructor(props) {
     super(props);
     this.handleSelectedDateUpdate = this.handleSelectedDateUpdate.bind(this);
-
-    this.handleDateChange = this.handleDateChange.bind(this);
-    this.handleDateChangeRaw = this.handleDateChangeRaw.bind(this);
     this.state = {
-      date: "2019-10-15", //moment().format(),
+      date: moment().format(),
     };
-  }
-
-  handleDateChange(event, date) {
-    this.setState({ date });
-  }
-
-  handleDateChangeRaw(event, date) {
-    this.setState({ date });
   }
 
   handleSelectedDateUpdate() {
@@ -31,11 +20,7 @@ class DatePickerDefault extends React.Component {
       <div>
         <DatePicker
           name="date-input"
-          // defaultValue={this.state.date}
-          // selectedDate={this.state.date}
-          value={this.state.date}
-          onChange={this.handleDateChange}
-          onChangeRaw={this.handleDateChangeRaw}
+          selectedDate={this.state.date}
         />
         <button type="button" onClick={this.handleSelectedDateUpdate}>Update selected date</button>
       </div>