+++ packages/terra-date-picker/tests/wdio/date-picker-spec.js
@@ -349,4 +349,48 @@ describe('Date Picker', () => {
       Terra.it.validatesElement({ axeRules: { rules: ignoredA11y } });
     });
   });
+
+  describe('Controlled DatePicker', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-date-picker/date-picker/date-picker-controlled');
+      // Removes the blinking cursor to prevent screenshot mismatches.
+      browser.execute('document.querySelector("input").style.caretColor = "transparent";');
+      browser.click('[class*="button"]');
+    });
+
+    Terra.it.matchesScreenshot('date picker opens', { selector: '[class="react-datepicker"]' });
+
+    it('Enters date value', () => {
+      browser.click('[class*="PopupOverlay"]');
+      browser.setValue('input[name="terra-date-controlled-date-picker"]', '07/01/2019');
+    });
+
+    Terra.it.matchesScreenshot('date input manually updated');
+
+    it('Open the date picker', () => {
+      browser.click('[class*="button"]');
+    });
+
+    Terra.it.matchesScreenshot('date picker updated', { selector: '[class="react-datepicker"]' });
+
+    it('Click button to reset date', () => {
+      browser.click('[class*="PopupOverlay"]');
+      browser.click('#reset-button');
+    });
+
+    Terra.it.matchesScreenshot('date input reset');
+
+    it('Open the date picker', () => {
+      browser.click('[class*="button"]');
+    });
+
+    Terra.it.matchesScreenshot('date picker reset', { selector: '[class="react-datepicker"]' });
+
+    it('Select date', () => {
+      browser.keys('ArrowDown');
+      browser.keys('Enter');
+    });
+
+    Terra.it.matchesScreenshot('date selected');
+  });
 });